third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-audio-element/video-loading-attr-lazy.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Video loading attribute should reflect the loading content attribute
  assert_equals: Video loading attribute should be 'lazy' when set to 'lazy' expected (string) "lazy" but got (undefined) undefined
Harness: the test ran to completion.
